Add style.css style.css on the main page by @waterfist #94
