+++ pages/main/index.tsx
@@ -3,11 +3,8 @@
 import Header from '@/components/Header';
 import Modal from '@/components/main/Modal';
 import { v4 as uuidv4 } from 'uuid';
-
 import { ChangeEvent, useCallback, useEffect, useRef, useState } from 'react';
-
 import styled from 'styled-components';
-import ModalLogin from '@/components/ModalLogin';
 import Seo from '@/components/Seo';
 import Chat from '@/components/chat/Chat';
 import { useInfiniteQuery } from 'react-query';
@@ -15,8 +12,6 @@ import { useBottomScrollListener } from 'react-bottom-scroll-listener';
 import { getInfiniteData, visibleReset } from '@/api';
 import Content from '@/components/main/Content';
 import { authService } from '@/firebase';
-import { signOut } from 'firebase/auth';
-import { customAlert } from '@/utils/alerts';
 import { useRouter } from 'next/router';
 import Search from '@/components/main/Search';
 import { CustomModal } from '@/components/common/CustomModal';
@@ -28,7 +23,6 @@ export default function Main() {
   const [isOpenModal, setOpenModal] = useState(false);
   const [chatToggle, setChatToggle] = useState(false);
   const [closeLoginModal, setCloseLoginModal] = useState(false);
-  const [isProfileOpen, setIsProfileOpen] = useState(false);
   const [currentUser, setCurrentUser] = useState(false);
   const [searchOption, setSearchOption] = useState('userName');
   const [searchValue, setSearchValue] = useState('');
@@ -49,20 +43,6 @@ export default function Main() {
     if (authService.currentUser) {
       setOpenModal(!isOpenModal);
     }
-  };
-  // 로그인 모달 창 버튼
-  const closeLoginModalButton = () => {
-    setCloseLoginModal(!closeLoginModal);
-  };
-
-  // 로그아웃
-  const logOut = () => {
-    signOut(authService).then(() => {
-      // Sign-out successful.
-      localStorage.clear();
-      setCurrentUser(false);
-      customAlert('로그아웃에 성공하였습니다!');
-    });
   };
 
   const onClickChatToggle = () => {
@@ -130,132 +110,127 @@ export default function Main() {
   }, [nowUser, router]);
 
   return (
-    <>
-      <div>
-        <Seo title="Home" />
+    <MainContainer>
+      <Seo title="Home" />
+      <MainHeaderdiv>
         <Header />
-        {/* 로그인, 로그아웃, 마이페이지 버튼 */}
-        {closeLoginModal && <ModalLogin closeModal={closeLoginModalButton} />}
-        <Profile onClick={() => setIsProfileOpen(!isProfileOpen)}>
-          {authService.currentUser?.photoURL ? (
-            <ProfileImg src={authService.currentUser?.photoURL} />
+        <div style={{ display: 'flex', gap: '10px', padding: '10px' }}>
+          <Categories value={selectCity} onChange={onChangeSelectCity}>
+            <option value="제주전체">제주전체</option>
+            <option value="제주시">제주시</option>
+            <option value="서귀포시">서귀포시</option>
+          </Categories>
+          {selectCity === '제주시' ? (
+            <Categories value={selectTown} onChange={onChangeSelectTown}>
+              <option value="">읍면동전체</option>
+              <option value="애월읍">애월읍</option>
+              <option value="남원읍">남원읍</option>
+            </Categories>
+          ) : selectCity === '서귀포시' ? (
+            <Categories value={selectTown} onChange={onChangeSelectTown}>
+              <option value="">읍면동전체</option>
+              <option value="표선면">표선면</option>
+              <option value="대정읍">대정읍</option>
+            </Categories>
           ) : (
-            <ProfileImg src="/profileicon.svg" />
+            <Categories value={selectTown} onChange={onChangeSelectTown}>
+              <option value="">읍면동전체</option>
+              <option value="표선면">표선면</option>
+              <option value="대정읍">대정읍</option>
+              <option value="애월읍">애월읍</option>
+              <option value="남원읍">남원읍</option>
+            </Categories>
           )}
-        </Profile>
-        {isProfileOpen === true ? (
-          <Menu>
-            {currentUser ? (
-              <MenuItem onClick={() => router.push('/mypage')}>
-                {' '}
-                마이페이지
-              </MenuItem>
-            ) : (
-              <MenuItem hidden onClick={() => router.push('/mypage')} />
-            )}
-            {currentUser ? (
-              <MenuItem onClick={logOut}>로그아웃</MenuItem>
-            ) : (
-              <MenuItem onClick={closeLoginModalButton}>로그인</MenuItem>
-            )}
-          </Menu>
-        ) : null}
-      </div>
-      {isOpenModal && (
-        <Modal
-          onClickToggleModal={onClickToggleModal}
-          setOpenModal={setOpenModal}
-        >
-          <div>children</div>
-        </Modal>
-      )}
-      <Search
-        searchOptionRef={searchOptionRef}
-        searchValue={searchValue}
-        onChangeSearchValue={onChangeSearchValue}
-      />
-      <div style={{ display: 'flex', gap: '10px', padding: '10px' }}>
-        <Categories value={selectCity} onChange={onChangeSelectCity}>
-          <option value="제주전체">제주전체</option>
-          <option value="제주시">제주시</option>
-          <option value="서귀포시">서귀포시</option>
-        </Categories>
-        {selectCity === '제주시' ? (
-          <Categories value={selectTown} onChange={onChangeSelectTown}>
-            <option value="">읍면동전체</option>
-            <option value="애월읍">애월읍</option>
-            <option value="남원읍">남원읍</option>
-          </Categories>
-        ) : selectCity === '서귀포시' ? (
-          <Categories value={selectTown} onChange={onChangeSelectTown}>
-            <option value="">읍면동전체</option>
-            <option value="표선면">표선면</option>
-            <option value="대정읍">대정읍</option>
-          </Categories>
-        ) : (
-          <Categories value={selectTown} onChange={onChangeSelectTown}>
-            <option value="">읍면동전체</option>
-            <option value="표선면">표선면</option>
-            <option value="대정읍">대정읍</option>
-            <option value="애월읍">애월읍</option>
-            <option value="남원읍">남원읍</option>
-          </Categories>
-        )}
-        <Categories>지역</Categories>
-        <Categories>팔로우</Categories>
-        <PostFormButton onClick={onClickToggleModal}>
-          게시물 작성
-        </PostFormButton>
-      </div>
-      <h1>{selectCity === 'undefined' ? '로딩중입니다.' : selectCity}</h1>
-      <div>
-        {/* 아래는 무한 스크롤 테스트 코드입니다. 차후, 메인페이지 디자인에 따라 바뀔 예정입니다. */}
-        {status === 'loading' ? (
-          <div>로딩중입니다.</div>
-        ) : status === 'error' ? (
-          <div>데이터를 불러오지 못했습니다.</div>
-        ) : (
-          <div>
-            <GridBox>
-              {data?.pages.map((data) =>
-                data?.map((item: any) => (
-                  <ItemBox key={uuidv4()}>
-                    <Content item={item} />
-                  </ItemBox>
-                ))
-              )}
-            </GridBox>
-          </div>
-        )}
+          <Categories>지역</Categories>
+          <Categories>팔로우</Categories>
+        </div>
 
-        {chatToggle ? <Chat /> : null}
-        <ChatToggleBtn onClick={onClickChatToggle}>
-          {chatToggle ? '닫기' : '열기'}
-        </ChatToggleBtn>
-        {/* <SearchPlace /> */}
-      </div>
-      <MapModalBtn onClick={onClickToggleMapModal}>지도열기</MapModalBtn>;
-      {isModalActive ? (
-        <CustomModal
-          modal={isModalActive}
-          setModal={setIsModalActive}
-          width="1200"
-          height="700"
-          element={<ModalMaps />}
+        <Search
+          searchOptionRef={searchOptionRef}
+          searchValue={searchValue}
+          onChangeSearchValue={onChangeSearchValue}
         />
-      ) : (
-        ''
-      )}
-    </>
+        <PostFormButton onClick={onClickToggleModal}>
+          + 나의 스팟 추가
+        </PostFormButton>
+      </MainHeaderdiv>
+      <MainBodydiv>
+        {isOpenModal && (
+          <Modal
+            onClickToggleModal={onClickToggleModal}
+            setOpenModal={setOpenModal}
+          >
+            <div>children</div>
+          </Modal>
+        )}
+        <h1>{selectCity === 'undefined' ? '로딩중입니다.' : selectCity}</h1>
+        <div>
+          {/* 아래는 무한 스크롤 테스트 코드입니다. 차후, 메인페이지 디자인에 따라 바뀔 예정입니다. */}
+          {status === 'loading' ? (
+            <div>로딩중입니다.</div>
+          ) : status === 'error' ? (
+            <div>데이터를 불러오지 못했습니다.</div>
+          ) : (
+            <div>
+              <GridBox>
+                {data?.pages.map((data) =>
+                  data?.map((item: any) => (
+                    <ItemBox key={uuidv4()}>
+                      <Content item={item} />
+                    </ItemBox>
+                  ))
+                )}
+              </GridBox>
+            </div>
+          )}
+
+          {chatToggle ? <Chat /> : null}
+          <ChatToggleBtn onClick={onClickChatToggle}>
+            {chatToggle ? '닫기' : '열기'}
+          </ChatToggleBtn>
+          {/* <SearchPlace /> */}
+        </div>
+        <MapModalBtn onClick={onClickToggleMapModal}>
+          지도에서 핀 보기
+        </MapModalBtn>
+
+        {isModalActive ? (
+          <CustomModal
+            modal={isModalActive}
+            setModal={setIsModalActive}
+            width="1200"
+            height="700"
+            element={<ModalMaps />}
+          />
+        ) : (
+          ''
+        )}
+      </MainBodydiv>
+    </MainContainer>
   );
 }
 
-const Categories = styled.select`
-  background-color: tomato;
-  width: 100px;
-  height: 40px;
+const MainContainer = styled.div``;
+
+const MainHeaderdiv = styled.div`
+  display: flex;
+  align-items: center;
+  height: 80px;
 `;
+
 const PostFormButton = styled.button`
+  border-radius: 20px;
+  color: cornflowerblue;
+  border: 1px solid cornflowerblue;
+  background-color: none;
+  width: 130px;
+  height: 25px;
+  cursor: pointer;
+`;
+
+const MainBodydiv = styled.div``;
+
+const Categories = styled.select`
   background-color: tomato;
   width: 100px;
   height: 40px;
@@ -268,13 +243,13 @@ const GridBox = styled.div`
   margin: 10px;
 `;
 const ItemBox = styled.div`
-  background-color: aqua;
+  background-color: pink;
   height: 250px;
   margin: 10px;
 `;
 const ChatToggleBtn = styled.button`
   position: fixed;
-  background-color: aqua;
+  background-color: cornflowerblue;
   left: 90%;
   top: 90%;
   border-radius: 50%;
@@ -282,48 +257,6 @@ const ChatToggleBtn = styled.button`
 
   width: 50px;
   height: 50px;
-`;
-const Profile = styled.div`
-  position: absolute;
-  top: 15px;
-  right: 20px;
-  z-index: 999;
-  width: 70px;
-  height: 70px;
-  border-radius: 50px;
-  background-color: white;
-  cursor: pointer;
-`;
-
-const ProfileImg = styled.img`
-  width: 70px;
-  height: 70px;
-  border-radius: 50px;
-  object-fit: cover;
-  position: fixed;
-  top: 10;
-  left: 80;
-  right: 20;
-`;
-
-const Menu = styled.div`
-  width: 100px;
-  height: 85px;
-  position: absolute;
-  top: 95px;
-  right: 20px;
-  background-color: orange;
-  border-radius: 5px;
-  box-shadow: 1px 1px 1px orange;
-  text-align: center;
-  color: white;
-  font-family: GmarketSans;
-
-  z-index: 6000;
-`;
-
-const MenuItem = styled.p`
-  cursor: pointer;
 `;
 
 const MapModalBtn = styled.button`
@@ -336,7 +269,7 @@ const MapModalBtn = styled.button`
   cursor: pointer;
   position: fixed;
   bottom: 10px;
-  width: 80px;
+  width: 100px;
   right: 50%;
   left: 50%;
 `;
