+++ pages/mypage/index.tsx
@@ -65,7 +65,7 @@ export default function Mypage() {
         <Seo title="My" />
         <h1>마이페이지입니다</h1>
       </MyTextDiv>
-      <Link href={'/'}>
+      <Link href={'//main?city=제주전체'}>
         {authService.currentUser ? (
           <LogoutButton onClick={logOut}>로그아웃</LogoutButton>
         ) : null}
